requirements.txt
# =============================================================================
# Mektep Scraper Dependencies
# =============================================================================

# Core Framework
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Migrate>=4.0,<5
Flask-Login>=0.6,<1
Flask-WTF>=1.2,<2
WTForms>=3.1,<4

# Database
# SQLite: built-in (for development)
# PostgreSQL: psycopg (for production)
psycopg[binary]>=3.1,<4

# Redis & Celery (async task queue)
redis>=5.0,<6
celery>=5.3,<6

# Web Scraping
playwright>=1.40,<2

# Document Generation
openpyxl>=3.1,<4
python-docx>=1.1,<2

# Security
cryptography>=42,<45

# Production Server
gunicorn>=21,<23;platform_system!="Windows"  # Unix only
waitress>=3.0,<4  # Windows compatible

# Configuration
python-dotenv>=1.0,<2

# AI Text Generation (optional)
openai>=1.0,<2